functions/requirements.txt
functions-framework==3.*
requests==2.28.2
google-auth==2.17.3
asyncio==3.4.3
aiohttp==3.8.4